Revert "add menu"
This reverts commit af3de8415d99bb0164c7d49c4785064932ceafc3.
