+++ src/app/auth-sign-up/auth-sign-up.page.scss
