+++ src/app/auth/auth.page.ts
@@ -1,9 +1,22 @@
 import { Component, OnInit } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { CommonModule, Location } from '@angular/common';
 import { FormsModule, FormGroup, FormControl, Validators, ReactiveFormsModule } from '@angular/forms';
 import { AuthService } from '../services/auth.service';
 import { Router, RouterLink } from '@angular/router';
-import { IonContent, IonHeader, IonTitle, IonToolbar, IonButton, IonItem, IonLabel, IonMenu, IonMenuButton, IonButtons, IonList } from '@ionic/angular/standalone';
+import { 
+  IonContent, 
+  IonHeader, 
+  IonTitle, 
+  IonToolbar, 
+  IonButton, 
+  IonItem, 
+  IonLabel, 
+  IonMenu, 
+  IonMenuButton, 
+  IonButtons, 
+  IonList, 
+  IonInput
+} from '@ionic/angular/standalone';
 
 import { CustomInputComponent } from "../components/custom-input/custom-input.component";
 @Component({
@@ -11,7 +24,10 @@ import { CustomInputComponent } from "../components/custom-input/custom-input.co
   templateUrl: './auth.page.html',
   styleUrls: ['./auth.page.scss'],
   standalone: true,
-  imports: [IonList, IonContent, 
+  imports: [
+    IonInput, 
+    IonList, 
+    IonContent, 
     IonHeader, 
     IonTitle, 
     IonToolbar, 
@@ -39,7 +55,11 @@ export class AuthPage implements OnInit {
     console.log(value);
   }
 
-  constructor(private authService: AuthService) { }
+  constructor(
+    private authService: AuthService, 
+    private router: Router,
+    private location:Location,
+  ) { }
 
   onSubmit() {
     console.log('Form Submitted', this.loginForm.value);
@@ -50,13 +70,14 @@ export class AuthPage implements OnInit {
         .then(result => {
           console.log("You are successfully logged in!", result);
           // Navigate to the desired route upon successful login
+          this.router.navigate(['/home']);
         })
         .catch(error => {
           console.error("Error logging in: ", error);
           // Provide user feedback about the login error
         });
     } else {
-      console.error('Form is not valid:', this.loginForm.errors);
+      console.error('Form is not valid:',  this.loginForm.errors);
       // Iterate over the form controls and set them as touched to trigger validation messages
       Object.keys(this.loginForm.controls).forEach(field => {
         const control = this.loginForm.get(field);
@@ -64,6 +85,11 @@ export class AuthPage implements OnInit {
       });
     }
   }
+
+  goBack() {
+    this.location.back();
+  }
+
   ngOnInit() {
   }
