+++ src/app/home/home.page.html
@@ -1,0 +1,6 @@
+<ion-header>
+  <ion-toolbar>
+      
+  </ion-toolbar>
+</ion-header>
+
