--- src/app/pages/auth/auth.page.spec.ts
@@ -1,17 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { AuthPage } from './auth.page';
-
-describe('AuthPage', () => {
-  let component: AuthPage;
-  let fixture: ComponentFixture<AuthPage>;
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(AuthPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
