--- src/app/pages/home/home.page.html
@@ -1,6 +1,0 @@
-<ion-content [fullscreen]="true">
-  <app-menu></app-menu>
-  <div class="container" id="container">
-    <app-item-explorer></app-item-explorer>
-  </div>
-</ion-content>
