--- src/app/pages/home/home.page.scss
@@ -1,3 +1,0 @@
-.container{
-    margin-top: 60px;
-}
