--- src/app/pages/profile/profile.page.scss
@@ -1,14 +1,0 @@
-.menu{
-  display: flex;
-  justify-content: center;
-  padding: 30px;
-
-}
-ion-menu-button {
-  color: var(--ion-color-primary);
-}
-ion-button {
-  display: flex;
-  justify-content: center;
-  padding: 10px;
-}
