+++ src/app/profile/profile.page.html
@@ -1,0 +1,76 @@
+<ion-header [translucent]="true">
+  <ion-toolbar>
+    <ion-buttons slot="start">
+      <ion-menu-button></ion-menu-button>
+    </ion-buttons>
+    <ion-title>Profile</ion-title>
+  </ion-toolbar>
+</ion-header>
+
+<ion-content>
+  <ion-header collapse="condense">
+    <ion-toolbar>
+    </ion-toolbar>
+
+  </ion-header>
+  <ion-card class="user-description-card">
+    <ion-card-content>
+      <ion-list lines="none">
+        <ion-item>
+          <ion-avatar slot="start">
+            <img src="/assets/icon/profile.jpg"/>
+          </ion-avatar>
+          <ion-label>
+            <h1>Lee Sang-hyeok</h1>
+            <p>Nombre</p>
+          </ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label>
+            <h2>(507) 6652-9849</h2>
+            <p>Numero de telefono</p>
+          </ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label>
+            <h2>faker&#64;t1.com</h2>
+            <p>Correo electronico</p>
+          </ion-label>
+        </ion-item>
+      </ion-list>
+    </ion-card-content>
+  </ion-card>
+
+  <ion-card class="user-options-card">
+    <ion-card-content>
+      <h2>Servicios</h2>
+        <ion-item>
+          <ion-label>
+           <ion-button>Tus pedidos</ion-button>
+          </ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label>
+           <ion-button>Cupones</ion-button>
+          </ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label>
+           <ion-button>Direcciones</ion-button>
+          </ion-label>
+        </ion-item>
+    </ion-card-content>
+  </ion-card>
+
+<ion-card class="user-services-card">
+  <ion-card-content>
+    <h2>Otros servicios</h2>
+      <ion-item>
+        <ion-label>
+         <ion-button>Soporte</ion-button>
+        </ion-label>
+      </ion-item>
+  </ion-card-content>
+</ion-card>
+
+</ion-content>
