+++ src/app/profile/profile.page.scss
@@ -1,0 +1,8 @@
+ion-menu-button {
+  color: var(--ion-color-primary);
+}
+ion-button {
+  display: flex;
+  justify-content: center;
+  padding: 10px;
+}
